datea_api/requirements/requirements.txt
Django==1.6.1
Pillow==2.2.2
South==0.8.4
django-extensions==1.2.5
django-tastypie==0.11.0
psycopg2==2.5.1
python-dateutil==2.2
python-memcached==1.53
python-mimeparse==0.1.4
six==1.4.1
sorl-thumbnail==11.12
wsgiref==0.1.2
